+++ packages/eslint-plugin-slds/src/v9/rules/no-important-tag.ts
@@ -1,9 +1,7 @@
 import { Rule } from 'eslint';
-import { rulesMetadata } from 'slds-shared-utils';
+import { rulesMetadata, isTargetProperty } from 'slds-shared-utils';
 
 const warningMsg = rulesMetadata['slds/no-important-tag'].warningMsg;
-
-// Remove isTargetProperty, just use true directly
 
 const rule: Rule.RuleModule = {
   meta: {
@@ -13,14 +11,28 @@ const rule: Rule.RuleModule = {
       recommended: false,
     },
     fixable: 'code',
-    schema: [],
+    schema: [
+      {
+        type: 'object',
+        properties: {
+          propertyTargets: {
+            type: 'array',
+            items: { type: 'string' }
+          }
+        },
+        additionalProperties: false
+      }
+    ],
   },
   create(context) {
+    const options = context.options[0] || {};
+    const propertyTargets = options.propertyTargets || [];
+    
     return {
       Declaration(node: any) {
         if (!context.filename.match(/\.(css|scss)$/)) return;
-        // Match all properties, as in stylelint
-        if (node.important /* && isTargetProperty(node.property, []) */) {
+        // Match properties based on targets, maintaining stylelint parity
+        if (node.important && isTargetProperty(node.property, propertyTargets)) {
           // Report at the property location
           const start = node.loc.start;
           const end = node.loc.end;
